+++ lib/components/BarChart/BarChart.js
@@ -6,26 +6,7 @@ import numeral from 'numeral';
 import { vds as designTokens } from '@/styles/build/js/design_tokens';
 import styles from './BarChart.css';
 
-const BAR_OFFSET = -45;
-const LABEL_OFFSET = 5;
-
-const CustomBarComponent = ({ x, y, width, height, label }) => (
-  <g transform="translate(0, 48)" className={styles.barContainer}>
-    <rect className={styles.barEdges} x={x} y={y + BAR_OFFSET} width={5} height={height} />
-    <rect className={styles.bar} x={x} y={y + BAR_OFFSET} width={width} height={height} />
-    <text dx="-70" y={y + height / 2 + BAR_OFFSET + LABEL_OFFSET} className={styles.barLabel}>
-      {label}
-    </text>
-  </g>
-);
-
-CustomBarComponent.propTypes = {
-  x: PropTypes.number.isRequired,
-  y: PropTypes.number.isRequired,
-  width: PropTypes.number.isRequired,
-  height: PropTypes.number.isRequired,
-  label: PropTypes.string.isRequired,
-};
+import CustomBar from './components/CustomBar/CustomBar';
 
 const BarChart = ({ data, yLabel, xLabel, yTicks, xTicks, onClick }) => {
   const dataToShow = data.sort((a, b) => a.value - b.value);
@@ -46,10 +27,10 @@ const BarChart = ({ data, yLabel, xLabel, yTicks, xTicks, onClick }) => {
   }
 
   if (!yTicks) {
-    axisLeft.renderTick = () => <></>;
+    axisLeft.renderTick = () => null;
   }
   if (!xTicks) {
-    axisBottom.renderTick = () => <></>;
+    axisBottom.renderTick = () => null;
     axisBottom.legendOffset = 10;
   }
 
@@ -77,12 +58,12 @@ const BarChart = ({ data, yLabel, xLabel, yTicks, xTicks, onClick }) => {
         layout="horizontal"
         indexBy="key"
         borderRadius="5"
-        padding=".2"
+        padding={0.2}
         layers={['grid', 'axes', 'bars', 'markers', 'legends', bottomAxisLine]}
         colors={designTokens.color.blue.lighter.value}
         labelTextColor={designTokens.color.blue.darker.value}
         enableGridY={false}
-        barComponent={CustomBarComponent}
+        barComponent={CustomBar}
         enableGridX
         onClick={onClick}
         isInteractive={false}
