+++ lib/components/BarChart/components/CustomBar/CustomBar.css
@@ -1,0 +1,28 @@
+.barContainer {
+  cursor: pointer;
+  transform: translateX(-20) !important;
+}
+
+.barContainer:hover {
+  opacity: 0.8;
+}
+
+.bar {
+  rx: 5;
+  ry: 5;
+  fill: var(--vds-color-blue-lighter);
+  stroke-width: 0;
+  stroke: var(--vds-color-blue-lighter);
+}
+
+.barEdges {
+  fill: var(--vds-color-blue-lighter);
+  stroke-width: 0;
+  stroke: var(--vds-color-blue-lighter);
+}
+
+.barLabel {
+  fill: var(--vds-color-blue-darker);
+  font-family: var(--vds-font-family-body);
+  font-size: var(--vds-font-size-sm);
+}
